pyproject.toml
[tool.poetry]
name = "proxychain-config-builder"
version = "1.0.0"
description = ""
authors = ["Opticore IT <opensource@opticoreit.com>"]

[tool.poetry.dependencies]
python = "^3.10"
click = "^8.0.4"
pysocks = "^1.7.1"
rich = "^13.7.0"

[tool.poetry.dev-dependencies]
pytest = "^5.4.1"
requests_mock = "^1.7.0"
black = "^19.10b0"
pylint = "^2.4.4"
pydocstyle = "^5.0.2"
bandit = "^1.6.2"
invoke = "^1.4.1"
toml = "0.10.1"
flake8 = "^3.8.3"

[tool.poetry.scripts]
proxychain-config-builder = 'proxychain_config_builder.cli:main'

[tool.black]
line-length = 120
include = '\.pyi?$'
exclude = '''
        /(
          \.git
        | \.tox
        | \.venv
        | env/
        | _build
        | build
        | dist
        | netbox/*
        )/
    '''

[tool.pylint.master]
ignore-paths = [
    ".*third_party.*",
    ".*tests.*",
]

[tool.pylint.messages_control]
# TODO: re-enable and fix these as time permits
# unused-import is already covered by flake8
disable=""",
    line-too-long,
    super-init-not-called,
    super-with-arguments,
    too-many-ancestors,
    too-many-arguments,
    too-many-boolean-expressions,
    too-many-branches,
    too-many-instance-attributes,
    too-many-lines,
    too-many-locals,
    too-many-nested-blocks,
    too-many-public-methods,
    too-many-return-statements,
    too-many-statements,
    wrong-import-order,
    wrong-import-position,
    """

[tool.pylint.miscellaneous]
# Don't flag TODO as a failure, let us commit with things that still need to be done in the code
notes = """,
    FIXME,
    XXX,
    """

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
